simsave.c compiles without warnings; fixed bug in force_full_save

+++ simsave.c
@@ -62,11 +62,19 @@ void handle_player_save()
       /* clear temp directory */
       if (!bootup_temp_clean_done)
       {
-	sprintf(command, "rm -f %s*", PLAYER_TEMP_DIR);
-	system(command);
-	sprintf(command, "rm -f %s*", BOX_TEMP_DIR);
-	system(command);
-	bootup_temp_clean_done = TRUE;
+        sprintf(command, "rm -f %s*", PLAYER_TEMP_DIR);
+        if ( system(command) == -1 )
+        {
+            bugf("handle_player_save: failed to execute command '%s'", command);
+            exit(1);
+        }
+        sprintf(command, "rm -f %s*", BOX_TEMP_DIR);
+        if ( system(command) == -1 )
+        {
+            bugf("handle_player_save: failed to execute command '%s'", command);
+            exit(1);
+        }
+        bootup_temp_clean_done = TRUE;
       }
     }
     break;
@@ -87,7 +95,7 @@ void handle_player_save()
        
        /* See if corresponding box in box_mf_list */
        sprintf(buf, "%s_box", mf->filename);
-       if ( box_mf = memfile_from_list(buf,box_mf_list) )
+       if ( (box_mf = memfile_from_list(buf,box_mf_list)) )
        {
 	   boxtemp = TRUE;
            if (!save_to_dir(box_mf, BOX_TEMP_DIR))
@@ -108,11 +116,19 @@ void handle_player_save()
 
   case SAVE_STATE_TEMPCOPY:
     sprintf(command, "mv %s* %s", PLAYER_TEMP_DIR, PLAYER_DIR);
-    system(command);
+    if ( system(command) == -1 )
+    {
+        bugf("handle_player_save: failed to execute command '%s'", command);
+        exit(1);
+    }
     if (boxtemp)
     {
       sprintf(command, "mv %s* %s", BOX_TEMP_DIR, BOX_DIR);
-      system(command);
+      if ( system(command) == -1 )
+      {
+          bugf("handle_player_save: failed to execute command '%s'", command);
+          exit(1);
+      }
       boxtemp=FALSE;
     }
 
@@ -144,7 +160,7 @@ void force_full_save()
    log_string("force_full_save: start");
 #endif
   if (nosave)
-    player_save_state == SAVE_STATE_SIMSAVE;
+    player_save_state = SAVE_STATE_SIMSAVE;
   else
     /* flush pending saves */
     while (player_save_state != SAVE_STATE_SIMSAVE)
@@ -189,10 +205,8 @@ void final_player_save()
 void sim_save_to_mem()
 {
   MEMFILE *mf;
-  MEMFILE *old_mf;
   //DESCRIPTOR_DATA *d;
   CHAR_DATA *ch;
-  char bug_buf[MSL];
 #if defined(SIM_DEBUG)
    log_string("sim_save_to_mem: start");
 #endif
@@ -221,13 +235,7 @@ void sim_save_to_mem()
 	  return;
       }
       
-      /* make sure player not already in save list */
-      if (remove_from_save_list(mf->filename))
-      {
-/*	  sprintf(bug_buf, "sim_save_to_mem: file <%s> already in player_save_list",
-		  mf->filename);
-	  bug(bug_buf, 0);*/
-      }//no longer a bug
+      remove_from_save_list(mf->filename);
       /* add pfile to player_save_list */
       mf->next = player_save_list;
       player_save_list = mf;
@@ -394,17 +402,16 @@ bool load_storage_boxes(CHAR_DATA *ch )
   char filename[MAX_INPUT_LENGTH];
   char bug_buf[MSL];
   bool found_in_mem;
-  OBJ_DATA *obj;
   sh_int i;
 
  if (IS_NPC(ch))
  {
      bugf("load_storage_boxes called on NPC",0);
-     return;
+     return FALSE;
  }
 
   if (ch->pcdata->storage_boxes<1)
-    return;
+    return FALSE;
 
   send_to_char("As you enter, an employee brings in your boxes and sets them before you.\n\r",ch);
   for (i=1;i<=ch->pcdata->storage_boxes;i++)
